added Tests
Added a project that generates a cpp file that includes the generated `embedded.h` header and extracts the embedded files and writes them on disk.
Also added a project that compiles the generated cpp file.
And also added a unit test project that compares the MD5 checksums of the original and extracted files.